+++ Server/ServerCore/Session.h
@@ -18,10 +18,10 @@ public:
 	SOCKET GetSocket();
 	sockaddr_in& GetAddress();
 	void SetAddress(sockaddr_in& addr);
-	void SetSessionId(uint32 id);
-	const int GetSessionId() const;
-	void SetDbId(uint32 id);
-	const int GetDbId() const;
+	void SetSessionId(uint64 id);
+	const uint64 GetSessionId() const;
+	void SetDbId(uint64 id);
+	const uint64 GetDbId() const;
 
 public:
 	virtual HANDLE GetHandle() override;
@@ -62,8 +62,8 @@ protected:
 	SOCKET _socket = INVALID_SOCKET;
 	sockaddr_in _addr = {};
 	RecvBuffer _recvBuffer;
-	uint32 _sessionId;
-	uint32 _dbId;
+	uint64 _sessionId;
+	uint64 _dbId;
 
 private:
 	RecvEvent* _recvEvent;
